make campaigns icon flush with the left edge of other content
mdi-react icons all contain built-in spacing of (generally) 3px on all sides (in the 24x24 px reference size). When the icon is used in the title of a page, such as "(Campaigns icon) alice / campaign-name", this causes the campaigns icon to be indented more than the content below it, which looks bad.